+++ packages/koot/core/webpack/config/transform-config-server.js
@@ -151,7 +151,8 @@ module.exports = async (kootBuildConfig = {}) => {
             warnings: true,
             // our additional options
             moduleTrace: true,
-            errorDetails: true
+            errorDetails: true,
+            performance: false,
         })
     }